+++ changeLog.md
@@ -20,3 +20,4 @@
 - 【filters】 增加 showReset 控制重置按钮显示
 - 【DynamicTable】 增加多级表头渲染，支持无限嵌套
 - 【DynamicTable】 可设置请求方式，避免 TableColumn 缓存
+- 【filters】 增加搜索按钮 loading 方法